RavenDB-20503 Change test class to RavenTestBase
